refactor: rename localStorage util
